pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "steca-charger-py"
version = "0.1.0"
description = "Python client for the eSystems White Label Wallbox REST API"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Andreas Ertel", email = "andreas.ertel@kontron.com"}
]
keywords = ["wallbox", "charger", "steca", "esystems", "ev-charging", "api-client"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Home Automation",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/Kontron-eSystems/StecaChargerPy"
Repository = "https://github.com/Kontron-eSystems/StecaChargerPy"
Issues = "https://github.com/Kontron-eSystems/StecaChargerPy/issues"

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "requests-mock>=1.11.0",
  "PyYAML>=6.0.0",
  "types-PyYAML>=6.0.0",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools]
packages = ["StecaChargerPy"]
include-package-data = true
